CONN-1355:Use EntityExceptionBuilder to create Errorresponse
